Bugfix for Non-Initialized Variable
nslookup would crash when no results were returned after the previous
fix for dnslib usage.  This was due to addr not being initialized and
then attempted to be free'd if the query failed.  As the variable is no
longer required it was removed.

Minor netshares update to include the computer targeted in the output