ldapsearch: use DC hostname instead of IP address for LDAP connection
Previously, if no hostname was explictily specified, a domain controller
was automatically discovered and the LDAP connection was established
based on the resolved IP address.

This prevents the use of Kerberos for authentication which is problematic
in hardened environments or when using members of the Protected Users
group for which NTLM is disabled.